requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
streamlit>=1.25.0
gunicorn>=20.1.0
xlrd>=2.0.1
